Start work on documenting example
